Adding automated section
Continuously reading from cmd.txt file for commands to execute, and
execute commands when there are new ones.

+++ MarsEnvironmentSimulator/src/map/Rover.java
@@ -6,7 +6,7 @@ import org.lwjgl.util.glu.Sphere;
 
 public class Rover {
 	double x, y, z;
-	int roverAngle, shoulderHorizontalAngle, shoulderVerticalAngle, elbowAngle, wristAngle, cameraAngle, cameraRotationAngle, wheelAngle;
+	double roverAngle, shoulderHorizontalAngle, shoulderVerticalAngle, elbowAngle, wristAngle, cameraAngle, cameraRotationAngle, wheelAngle;
 	boolean cameraOn;
 	Cylinder cylinder;
 	Sphere joint;
